Algebra

Description

Documentation

class Group a where

Defines a Group. = Group Laws (See Wikipedia Article) (1) Closure: for a,b in G, a # b in G (2) Associativity: for a,b,c in G, (a c == a c) (3) Identity element: There exists e in G such that e e == a for all a in G (4) Inverse element: for a in G, there exists b in G such that a a == e

Minimal complete definition

(#), (/#), identity

Methods

(#) :: a -> a -> a

The group operation

(/#) :: a -> a -> a

The inverse of the group operation

identity :: a

The identity element of the group

inverse :: a -> a

The inverse of an element in the group